FIX linting issues
